This is okay

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -3,4 +3,5 @@
 exec { 'pkill':
   command  => 'pkill killmenow',
   provider => 'shell',
+ onlyif  => 'pgrep killmenow',
 }
